Converted Iframe and Stream middlewares to components.
