+++ components/stream/index.js
@@ -1,0 +1,6 @@
+function Iframe(server) {
+    const { handleStreamRequest } = require("./controller");
+    server.use(`/stream/:keySSI/*`, handleStreamRequest);
+}
+
+module.exports = Iframe;
