+++ components/stream/worker-script.js
