+++ index.js
@@ -25,6 +25,8 @@ const CHECK_FOR_RESTART_COMMAND_FILE_INTERVAL = 500;
 	require('./components/mqHub');
 	require('./components/enclave');
 	require('./components/secrets');
+	require('./components/iframe');
+	require('./components/stream');
 	//end
 })();
 
@@ -176,11 +178,9 @@ function HttpServer({ listeningPort, rootFolder, sslConfig, dynamicPort, restart
 			const LoggerMiddleware = require('./middlewares/logger');
 			const AuthorisationMiddleware = require('./middlewares/authorisation');
 			const OAuth = require('./middlewares/oauth');
-			const IframeHandlerMiddleware = require('./middlewares/iframeHandler');
 			const ResponseHeaderMiddleware = require('./middlewares/responseHeader');
 			const genericErrorMiddleware = require('./middlewares/genericErrorMiddleware');
 			const requestEnhancements = require('./middlewares/requestEnhancements');
-			const StreamHandlerMiddleware = require('./middlewares/streamHandler');
 
 			if(conf.enableRequestLogger) {
 				new LoggerMiddleware(server);
@@ -198,17 +198,10 @@ function HttpServer({ listeningPort, rootFolder, sslConfig, dynamicPort, restart
 			if(conf.responseHeaders){
 				new ResponseHeaderMiddleware(server);
 			}
-            if(conf.iframeHandlerDsuBootPath) {
-                new IframeHandlerMiddleware(server);
-            }
             if(conf.enableInstallationDetails) {
                 const enableInstallationDetails = require("./components/installation-details");
                 enableInstallationDetails(server);
             }
-            if(conf.enableStreamHandler) {
-                new StreamHandlerMiddleware(server);
-            }
-
         }
 
         function addComponent(componentName, componentConfig) {
